backend/requirements.txt
ffmpeg-python
openai
openai-whisper
pydantic-settings
